backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# PyTorch & Deep Learning
torch==2.2.0
torchvision==0.17.0

# Image Processing
pillow==10.2.0
opencv-python==4.9.0.80

# Data & Utilities
numpy<2.0.0
pandas==2.2.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1

# Validation & Config
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging
python-json-logger==2.0.7

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
